lie algebra split ceres related
