+++ src/aliceVision/geometry/CMakeLists.txt
@@ -5,6 +5,7 @@ set(geometry_files_headers
     Pose3.hpp
     rigidTransformation3D.hpp
     Similarity3.hpp
+    lie.hpp
 )
 
 # Sources
