+++ src/aliceVision/sfm/liealgebra.hpp
@@ -1,5 +1,12 @@
+// This file is part of the AliceVision project.
+// Copyright (c) 2023 AliceVision contributors.
+// This Source Code Form is subject to the terms of the Mozilla Public License,
+// v. 2.0. If a copy of the MPL was not distributed with this file,
+// You can obtain one at https://mozilla.org/MPL/2.0/.
+
 #pragma once
 
+#include <aliceVision/geometry/lie.hpp>
 #include <Eigen/Dense>
 #include <unsupported/Eigen/KroneckerProduct>
 #include <aliceVision/utils/CeresUtils.hpp>
@@ -7,85 +14,6 @@
 
 namespace aliceVision {
 namespace SO3 {
-
-using Matrix = Eigen::Matrix<double, 3, 3, Eigen::RowMajor>;
-
-/**
-Compute the skew symmetric matrix of the given vector 3d
-@param int the 3d vector
-@return a skew symmetric matrix
-*/
-inline Eigen::Matrix3d skew(const Eigen::Vector3d & in) {
-    Eigen::Matrix3d ret;
-
-    ret.fill(0);
-
-    ret(0, 1) = -in(2);
-    ret(1, 0) = in(2);
-    ret(0, 2) = in(1);
-    ret(2, 0) = -in(1);
-    ret(1, 2) = -in(0);
-    ret(2, 1) = in(0);
-
-    return ret;
-}
-
-/**
-Compute the exponential map of the given algebra on the group
-@param algebra the 3d vector
-@return a 3*3 SO(3) matrix
-*/
-inline Eigen::Matrix3d expm(const Eigen::Vector3d & algebra) {
-    double angle = algebra.norm();
-
-    if (angle < std::numeric_limits<double>::epsilon()) {
-        return Eigen::Matrix3d::Identity();
-    }
-
-    Eigen::Matrix3d omega = skew(algebra);
-
-    Eigen::Matrix3d ret;
-    ret = Eigen::Matrix3d::Identity() + (sin(angle) / angle) * omega + ((1.0 - cos(angle)) / (angle * angle)) * omega * omega;
-
-    return ret;
-}
-
-/**
-Compute the algebra related to a given rotation matrix
-@param R the input rotation matrix
-@return the algebra
-*/
-inline Eigen::Vector3d logm(const Eigen::Matrix3d & R) {
-
-    Eigen::Vector3d ret;
-
-    double p1 = R(2, 1) - R(1, 2);
-    double p2 = R(0, 2) - R(2, 0);
-    double p3 = R(1, 0) - R(0, 1);
-
-    double costheta = (R.trace() - 1.0) / 2.0;
-    if (costheta < -1.0) {
-        costheta = -1.0;
-    }
-
-    if (costheta > 1.0) {
-        costheta = 1.0;
-    }
-
-    if (1.0 - costheta < 1e-24) {
-        ret.fill(0);
-        return ret;
-    }
-
-    double theta = acos(costheta);
-    double scale = theta / (2.0 * sin(theta));
-
-    ret(0) = scale * p1;
-    ret(1) = scale * p2;
-    ret(2) = scale * p3;
-
-    return ret;
-}
 
 /**
 Compute the jacobian of the logarithm wrt changes in the rotation matrix values
@@ -201,41 +129,9 @@ class Manifold : public utils::CeresManifold {
   int TangentSize() const override { return 3; }
 };
 
-}
+}//namespace SO3 
 
 namespace SE3 {
-
-using Matrix = Eigen::Matrix<double, 4, 4, Eigen::RowMajor>;
-
-/**
-Compute the exponential map of the given algebra on the group
-@param algebra the 6d vector
-@return a 4*4 SE(3) matrix
-*/
-inline Eigen::Matrix4d expm(const Eigen::Matrix<double, 6, 1> & algebra){
-
-  Eigen::Matrix4d ret;
-  ret.setIdentity();
-
-  Eigen::Vector3d vecR = algebra.block<3, 1>(0, 0);
-  Eigen::Vector3d vecT = algebra.block<3, 1>(3, 0);
-
-
-  double angle = vecR.norm();
-  if (angle < std::numeric_limits<double>::epsilon()) {
-    ret.setIdentity();
-    ret.block<3, 1>(0, 3) = vecT;
-    return ret;
-  }
-
-  Eigen::Matrix3d omega = SO3::skew(vecR);
-  Eigen::Matrix3d V = Eigen::Matrix3d::Identity() + ((1.0 - cos(angle)) / (angle*angle)) * omega + ((angle - sin(angle)) / (angle*angle*angle)) * omega * omega;
-
-  ret.block<3, 3>(0, 0) = SO3::expm(vecR);
-  ret.block<3, 1>(0, 3) = V * vecT;
-
-  return ret;
-}
 
 
 class ManifoldLeft : public utils::CeresManifold {
@@ -379,30 +275,11 @@ private:
   bool _refineRotation;
   bool _refineTranslation;
 };
-}
+
+} //namespace SE3 
 
 
 namespace SO2 {
-
-using Matrix = Eigen::Matrix<double, 2, 2, Eigen::RowMajor>;
-
-/**
-Compute the exponential map of the given algebra on the group
-@param algebra the 1d vector
-@return a 2*2 S0(2) matrix
-*/
-inline Eigen::Matrix2d expm(double algebra){
-
-  Eigen::Matrix2d ret;
-  
-  ret(0, 0) = cos(algebra);
-  ret(0, 1) = -sin(algebra);
-  ret(1, 0) = sin(algebra);
-  ret(1, 1) = cos(algebra);
-
-  return ret;
-}
-
 
 class Manifold : public utils::CeresManifold {
 public:
@@ -447,6 +324,6 @@ public:
   }
 };
 
-}
+} //namespace SO2 
 
-}
+} //namespace aliceVision
